Refactor the vertex buffer param into draw()
If the vertex buffer's underlying data store is changed (via a call to
glBufferData), glVertexAttribPointer calls become invalid. This was a
soundness hole, so the API now prevents it.

It does leave room for a future optimization, where a shader checks to
see if the provided vertex buffer is the same one as last time, and
vertex buffers check if they have dumped their underlying data store.